Fix frozen slider before interaction
