+++ frontend/src/Containers/HomePage/HomePage.js
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
 import AboutUsSection from "./Sections/AboutUsSection";
 import SubheadingDivider from "Components/SubheadingDivider/SubheadingDivider";
 import CarouselGallery from "Components/Carousel/CarouselGallery";
@@ -8,9 +8,24 @@ import GallerySection from "./Sections/GallerySection";
 import WelcomeSection from "./Sections/WelcomeSection";
 
 const HomePage = ({ api }) => {
+  const [images, setImages] = useState([]);
+
+  useEffect(() => {
+    // Fetch cakes data from the API
+    const fetchImages = async () => {
+      try {
+        const response = await api.get("api/slider-images/");
+        setImages(response.data);
+      } catch (error) {
+        console.error("Error fetching cakes:", error);
+      }
+    };
+    fetchImages();
+  }, [api]);
+
   return (
     <>
-      <CarouselGallery api={api} />
+      {images.length > 0 && <CarouselGallery images={images} />}
       <WelcomeSection api={api} />
       <GallerySection api={api} />
       {/* <SubheadingDivider subheadingText="Cakes for all occasions" /> */}
